DrawDistIncrease: allows drawing objects from further away, heavily reducing pop-in
The way this works is kinda by making the game think your car is further along the track than it actually is (it's a bit more involved than that, but that's mostly what it boils down to)

Unfortunately the devs added some LOD models in the later parts of stages to simulate the beginning parts, in the stages where you can get a peek at the earlier sections.

So if the draw distance is pushed too far, those LOD versions of the beginning parts can end up drawing on top of the actual non-LOD version, sometimes resulting in some ugly mass appearing on the track, or the track section getting replaced entirely with the LOD version.

Pushing it a small amount (DistIncrease = 16 or so) should help with some of the worst pop-ins without having the LOD issue, but going further can cause problems, hopefully can add something to filter out the LOD models from being drawn soon...